# Backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API powering ScreenScore backend system.",
        "title": "ScreenScore Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/v1",
    "paths": {
        "/groups/create/{groupName}": {
            "post": {
                "description": "Creates a new group and returns invite code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Create new group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name (URL parameter)",
                        "name": "groupName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.GroupData"
                        }
                    },
                    "400": {
                        "description": "Group name malformed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/groups/join/{inviteCode}": {
            "post": {
                "description": "Joins a group and returns group id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Join group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invite code (URL parameter)",
                        "name": "inviteCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invite code malformed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/groups/kick": {
            "post": {
                "description": "Anyone can kick another group member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Kick user",
                "parameters": [
                    {
                        "description": "Payload with user to kick and group id",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.KickUserPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "User with given id is not a member of the group",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/groups/leave": {
            "post": {
                "description": "Leaves a group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Leave group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group id (URL parameter)",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Checks if the server is active.",
                "tags": [
                    "utils"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/stats/get-group-stats/{groupID}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Retrieves screentime data for all group memebers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group id (URL parameter)",
                        "name": "groupID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.GroupStats"
                            }
                        }
                    },
                    "403": {
                        "description": "User with given id is not a member of the group",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/stats/sync-stats": {
            "post": {
                "description": "Updates or adds a new screen time record for a specific device.\nIncludes logic to prevent \"cheating\" (time from future, screen time increasing faster than real time, etc.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Synchronize user screen time",
                "parameters": [
                    {
                        "description": "User screen time data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UserStatsPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "database updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Validation error (Future time, time travel, invalid increments)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Creates a new user account, hashes the password, and links the initial device with a push token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data (email, username, password, device_id, push_token)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UserPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "schema": {
                            "type": "nil"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error during hashing or database insert",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.GroupData": {
            "type": "object",
            "properties": {
                "invite_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "main.KickUserPayload": {
            "type": "object",
            "required": [
                "group_id",
                "user_to_kick_id"
            ],
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "user_to_kick_id": {
                    "type": "integer"
                }
            }
        },
        "main.UserPayload": {
            "type": "object",
            "required": [
                "device_id",
                "email",
                "password",
                "username"
            ],
            "properties": {
                "device_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 3
                },
                "push_notification_token": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "main.UserStatsPayload": {
            "type": "object",
            "required": [
                "device_id",
                "recorded_at",
                "screen_time"
            ],
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "recorded_at": {
                    "type": "string",
                    "example": "2026-03-17T12:00:00Z"
                },
                "screen_time": {
                    "type": "integer"
                }
            }
        },
        "storage.GroupStats": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "recorded_at": {
                    "type": "string"
                },
                "screen_time": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}